nm dns: Do not use current interface which is unmanaged/external managed
When search interface from current state, we should not use `unmanaged`
or `external managed` interface for storing DNS information.

Integration test case included.
