+++ rust/src/lib/nm/gen_conf.rs
@@ -26,7 +26,7 @@ pub(crate) fn nm_gen_conf(
     let mut merged_state = merged_state.clone();
     store_route_config(&mut merged_state)?;
     store_route_rule_config(&mut merged_state)?;
-    store_dns_config_to_iface(&mut merged_state)?;
+    store_dns_config_to_iface(&mut merged_state, &[], &[])?;
 
     let nm_conns = perpare_nm_conns(
         &merged_state,
