+++ rust/src/lib/nm/nm_dbus/query_apply/mod.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod device;
