+++ rust/src/lib/nm/settings/connection.rs
@@ -49,6 +49,21 @@ pub(crate) const NM_SETTING_USER_SPACES: [&str; 2] = [
     NM_SETTING_OVS_PORT_SETTING_NAME,
 ];
 
+pub(crate) const SUPPORTED_NM_KERNEL_IFACE_TYPES: [&str; 12] = [
+    NM_SETTING_WIRED_SETTING_NAME,
+    NM_SETTING_VETH_SETTING_NAME,
+    NM_SETTING_BOND_SETTING_NAME,
+    NM_SETTING_DUMMY_SETTING_NAME,
+    NM_SETTING_BRIDGE_SETTING_NAME,
+    NM_SETTING_OVS_IFACE_SETTING_NAME,
+    NM_SETTING_VRF_SETTING_NAME,
+    NM_SETTING_VLAN_SETTING_NAME,
+    NM_SETTING_VXLAN_SETTING_NAME,
+    NM_SETTING_MACVLAN_SETTING_NAME,
+    NM_SETTING_LOOPBACK_SETTING_NAME,
+    NM_SETTING_INFINIBAND_SETTING_NAME,
+];
+
 pub(crate) fn iface_to_nm_connections(
     merged_iface: &MergedInterface,
     merged_state: &MergedNetworkState,
