+++ rust/src/lib/nm/settings/mod.rs
@@ -23,7 +23,9 @@ mod vrf;
 mod vxlan;
 mod wired;
 
-pub(crate) use self::connection::{get_exist_profile, iface_to_nm_connections};
+pub(crate) use self::connection::{
+    get_exist_profile, iface_to_nm_connections, SUPPORTED_NM_KERNEL_IFACE_TYPES,
+};
 #[cfg(feature = "query_apply")]
 pub(crate) use self::connection::{
     iface_type_to_nm, NM_SETTING_BOND_SETTING_NAME,
